+++ src/domains/data/index.ts
@@ -5,3 +5,5 @@ export type DomainStatus = 'validating' | 'valid' | 'invalid';
 export interface Domain extends ShlinkDomain {
   status: DomainStatus;
 }
+
+export const DEFAULT_DOMAIN = 'DEFAULT';