+++ src/utils/features.ts
@@ -13,6 +13,7 @@ const supportedFeatures = {
   urlValidation: { maxVersion: '3.*.*' },
   ipRedirectCondition: { minVersion: '4.2.0' },
   geolocationRedirectCondition: { minVersion: '4.3.0' },
+  filterShortUrlsByDomain: { minVersion: '4.3.0' },
 } as const satisfies Record<string, Versions>;
 
 Object.freeze(supportedFeatures);
@@ -35,6 +36,7 @@ const getFeaturesForVersion = (serverVersion: SemVerOrLatest): Record<Feature, b
   urlValidation: isFeatureEnabledForVersion('urlValidation', serverVersion),
   ipRedirectCondition: isFeatureEnabledForVersion('ipRedirectCondition', serverVersion),
   geolocationRedirectCondition: isFeatureEnabledForVersion('geolocationRedirectCondition', serverVersion),
+  filterShortUrlsByDomain: isFeatureEnabledForVersion('filterShortUrlsByDomain', serverVersion),
 });
 
 const FeaturesContext = createContext(getFeaturesForVersion('0.0.0'));